requirements.txt
Flask==2.0.1
requests==2.25.1
Jinja2==2.11.3
Werkzeug==2.0.0
MarkupSafe==1.1.1
urllib3==1.26.4
cryptography==3.3.2
PyYAML==5.3.1
SQLAlchemy==1.3.23
Pillow==8.1.2
django==2.2.0
numpy==1.19.5